# public/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Content management system",
    "summary": "Built on Express with TypeScript.",
    "contact": {
      "name": "Repo",
      "url": "https://github.com/chrjl/cms-express-ts"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/posts": {
      "get": {
        "tags": ["posts", "keywords"],
        "summary": "Get a list of all posts. Optionally filter based on keyword tag.",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": {} } },
          "404": { "description": "Not Found" }
        },
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "explode": true
          }
        ]
      },
      "post": {
        "tags": ["posts"],
        "summary": "Create a new post.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "My new post",
                "description": "A brief description of what my new post is about"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/posts/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": "true",
          "schema": {
            "type": "integer"
          },
          "default": "1"
        }
      ],
      "get": {
        "tags": ["posts"],
        "summary": "Get content and all metadata of a single post.",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": {} } },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "tags": ["posts"],
        "summary": "Update content and/or metadata of a single post."
      },
      "patch": {
        "tags": ["posts"],
        "summary": "Update only metadata of a single post."
      },
      "delete": {
        "tags": ["posts"],
        "summary": "Delete a single post.",
        "responses": {
          "204": { "description": "No Content" }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "default": "9999"
          }
        ]
      }
    },
    "/posts/{id}/keywords": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": "true",
          "schema": {
            "type": "integer"
          },
          "default": "1"
        }
      ],
      "get": {
        "tags": ["keywords"],
        "summary": "Get the list of keywords associated with a single post.",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": {} } }
        }
      }
    },
    "/keywords": {
      "get": {
        "tags": ["keywords"],
        "summary": "Get a list of all keywords.",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": {} } }
        }
      }
    }
  }
}